fix(core): prevent duplicate prefixes when using outputStyle stream
