+++ package.json
@@ -299,6 +299,7 @@
     "regenerator-runtime": "0.13.7",
     "send": "0.17.1",
     "string-width": "^4.2.3",
+    "strong-log-transformer": "^2.1.0",
     "tailwindcss": "3.1.8",
     "tslib": "^2.3.0",
     "weak-napi": "^2.0.2"
