+++ packages/nx/package.json
@@ -55,6 +55,7 @@
     "open": "^8.4.0",
     "semver": "7.3.4",
     "string-width": "^4.2.3",
+    "strong-log-transformer": "^2.1.0",
     "tar-stream": "~2.2.0",
     "tmp": "~0.2.1",
     "tsconfig-paths": "^3.9.0",
